search-engine-service/requirements.txt
# Core
python-dotenv==1.0.1
clickhouse-connect==0.8.5
pandas>=3.0.1
numpy==1.26.4
bottleneck>=1.4.2
numexpr>=2.10.2

# ML/DL
torch==2.5.1
transformers==4.46.2
peft==0.14.0
sentence-transformers==3.3.1

# Vector Search
faiss-cpu==1.9.0

# Baseline
rank-bm25==0.2.2
symspellpy==6.9.0

# UI
streamlit==1.40.1

# API
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3

# Training/Hyperparameter Tuning
optuna==4.1.0

# Utils
tqdm==4.67.1
scikit-learn==1.6.1
scipy>=1.10.0
matplotlib>=3.5.0
psutil>=5.9.0
pyarrow>=15.0.0

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-asyncio==0.24.0
pytest-mock==3.14.0
httpx==0.28.1
